Finished Google Search functionality
