+++ Config.properties
@@ -1,1 +1,1 @@
-browserType=
+browser=chrome
