+++ src/test/java/com/cydeo/runners/CukesRunner.java
@@ -8,7 +8,8 @@ import org.junit.runner.RunWith;
 @CucumberOptions(
         features = "src/test/resources/features",
         glue = "com/cydeo/step_definitions",
-        dryRun = false
+        dryRun = false,
+        tags = "@wip"
 )
 public class CukesRunner {
