--- src/test/java/com/cydeo/step_definitions/Google_StapDefinitions.java
@@ -1,23 +1,0 @@
-package com.cydeo.step_definitions;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-public class Google_StapDefinitions {
-
-    @Given("User is on Google home page")
-    public void user_is_on_google_home_page() {
-
-    }
-
-    @When("User searches for apple")
-    public void user_searches_for_apple() {
-
-    }
-
-    @Then("User should see apple in the title")
-    public void user_should_see_apple_in_the_title() {
-
-    }
-}
