+++ src/test/java/com/cydeo/step_definitions/Google_StepDefinitions.java
@@ -1,0 +1,47 @@
+package com.cydeo.step_definitions;
+
+import com.cydeo.pages.GoogleSearchPage;
+import com.cydeo.utilities.Driver;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+import org.junit.Assert;
+import org.openqa.selenium.Keys;
+
+public class Google_StepDefinitions {
+
+    @Given("User is on Google home page")
+    public void user_is_on_google_home_page() {
+
+        Driver.getDriver().get("https://www.google.com");
+
+    }
+
+    @When("User searches for apple")
+    public void user_searches_for_apple() {
+
+        GoogleSearchPage googleSearchPage = new GoogleSearchPage();
+        googleSearchPage.searchBar.sendKeys("apple" + Keys.ENTER);
+
+
+    }
+
+    @Then("User should see apple in the title")
+    public void user_should_see_apple_in_the_title() {
+
+        String actualTitle = Driver.getDriver().getTitle();
+        String expectedInTitle = "apple";
+
+        Assert.assertTrue(actualTitle.contains(expectedInTitle));
+
+    }
+
+    @Then("User should see title is Google")
+    public void user_should_see_title_is_google() {
+
+        String actualTitle = Driver.getDriver().getTitle();
+        String expectedTitle = "Google";
+
+        Assert.assertTrue(actualTitle.equals(expectedTitle));
+    }
+}
