+++ src/test/resources/features/googleSearch.feature
@@ -2,6 +2,11 @@ Feature: Google search functionality
   Agile Story: As s user, When I am on the google search page,
   I should be able to search whatever I want, and see relevant information
 
+  Scenario: Search page title verification
+    Given User is on Google home page
+    Then User should see title is Google
+
+  @wip
   Scenario: User search title verification
     Given User is on Google home page
     When User searches for apple
